moves worker and transforms dirs
