+++ src/backend/transforms/branch.ts
@@ -1,7 +1,7 @@
-import { getJiraId } from "../common/id";
+import { getJiraId } from "../../common/id";
 import { Context } from "probot/lib/context";
 import issueKeyParser from "jira-issue-key-parser";
-import { isEmpty } from "../common/isEmpty";
+import { isEmpty } from "../../common/isEmpty";
 
 async function getLastCommit(context, issueKeys) {
 	const {
