+++ src/backend/transforms/push.ts
@@ -1,11 +1,11 @@
-import { Subscription } from "../backend/models";
-import getJiraClient from "../jira/client";
+import { Subscription } from "../models";
+import getJiraClient from "../../jira/client";
 import issueKeyParser from "jira-issue-key-parser";
-import { isEmpty } from "../common/isEmpty";
+import { isEmpty } from "../../common/isEmpty";
 import { queues } from "../worker/main";
-import enhanceOctokit from "../config/enhance-octokit";
+import enhanceOctokit from "../../config/enhance-octokit";
 import { Application } from "probot";
-import { getLogger } from "../config/logger";
+import { getLogger } from "../../config/logger";
 
 // TODO: define better types for this file
