+++ src/backend/worker/app.ts
@@ -4,8 +4,8 @@
 
 import { Application, createProbot } from "probot";
 import { findPrivateKey } from "probot/lib/private-key";
-import healthcheck from "../frontend/healthcheck";
-import { overrideProbotLoggingMethods } from "../config/logger";
+import healthcheck from "../../frontend/healthcheck";
+import { overrideProbotLoggingMethods } from "../../config/logger";
 
 export const probot = createProbot({
 	id: Number(process.env.APP_ID),
