+++ src/backend/worker/metrics-job.ts
@@ -1,7 +1,7 @@
-import { metricHttpRequest } from "../config/metric-names";
-import statsd from "../config/statsd";
-import { Subscription } from "../backend/models";
-import { getLogger } from "../config/logger";
+import { metricHttpRequest } from "../../config/metric-names";
+import statsd from "../../config/statsd";
+import { Subscription } from "../models";
+import { getLogger } from "../../config/logger";
 
 const logger = getLogger("metrics-job");
