+++ src/github/deployment.ts
@@ -1,4 +1,4 @@
-import transformDeployment from "../transforms/deployment";
+import transformDeployment from "../backend/transforms/deployment";
 import { Context } from "probot/lib/context";
 
 export default async (context: Context, jiraClient): Promise<void> => {
